test: enable JS-Utils unit and integ
